Hoist GraphicSectorMap out of Map.py
